Introducing Channeled Spells, spells that trigger while you're channeling your Spellblade by shift right clicking once they're activated.
Introducing Nullifying Stance, a channeled spell that removes projectiles around you

Introducing Enderloper Stance, a channeled spell that attempts to teleport you behind whatever hits you in melee range. If it can't do it, it instead teleports you directly on the target.

Introducing Sentinels, mobs that target enemies that do not have Totemic Zeal as long as the Sentinel is within range of a Sentinel Totem (new)

Introducing Animate, a spell that is used on iron blocks to create Sentinels, or on swords on the ground in order to create animated blades that follow you and fight for you.

Introducing Blade Flurry, a channeled spell that conjures blades that spin around you in a sphere centered slightly in front of you, with you within it.

Reintroducing Vengeful Stance, a spell that hits enemies with your main hand when they hit you.

Fixed multiplayer crash related to attempting to load clientside class in dedicated server

Reverberating ray now has larger initial spread